pyproject.toml
[tool.poetry]
name = "django-p8-template"
version = "0.1.0"
description = ""
authors = ["xolmomin <xolmomin@gmail.com>"]
readme = "README.md"
packages = [{include = "django_p8_template"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
flake8 = "^6.0.0"
isort = "^5.12.0"
django-mptt = "^0.14.0"
django-ckeditor = "^6.5.1"
pillow = "^9.5.0"
fixtures = "^4.0.1"
pyaml = "^23.5.6"
pymal = "^0.6"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
